finish subtask 9 - modules
